Refactor fine-tuning job creation to include is_public flag
This commit modifies the `create_fine_tuning_job` method in the `FineTuningHandle` class to include an `is_public` flag. The flag is set to `True` by default and is used to determine whether the fine-tuning job should be publicly accessible. This change allows for more flexibility in controlling the visibility of fine-tuning jobs.

The `create_fine_tuning_job` method now accepts an additional parameter `is_public`, which defaults to `True`. The value of `is_public` is stored in the `is_public` column of the `fine_tuning_jobs` table in the database.

This commit also includes a database migration script that adds the `is_public` column to the `fine_tuning_jobs` table.

Lastly, the default language in the frontend has been changed from French to English in the `i18n.js` file. This ensures that the default language is set correctly.